% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dashboardCardsApp.R
\name{modB_dashboardCardsApp_server}
\alias{modB_dashboardCardsApp_server}
\title{Module B: NPS Server Function}
\usage{
modB_dashboardCardsApp_server(id)
}
\description{
Module B: NPS Server Function
}
