% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modB_dataAnalysisApp.R
\name{modB_dataAnalysisApp_server}
\alias{modB_dataAnalysisApp_server}
\title{dataAnalysis Server Functions}
\usage{
modB_dataAnalysisApp_server(id, dataset_selection, processed_data)
}
\arguments{
\item{dataset_selection}{Reactive containing selected dataset}

\item{processed_data}{Reactive containing processed count data}
}
\description{
dataAnalysis Server Functions
}
