% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_interactiveDashboardApp.R
\name{modB_interactiveDashboardApp_server}
\alias{modB_interactiveDashboardApp_server}
\title{Module B: NPS Server Function}
\usage{
modB_interactiveDashboardApp_server(id, module_states)
}
\description{
Module B: NPS Server Function
}
