% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_interactiveDashboardApp.R
\name{modB_interactiveDashboardApp_ui}
\alias{modB_interactiveDashboardApp_ui}
\title{Module B: NPS UI Function}
\usage{
modB_interactiveDashboardApp_ui(id)
}
\description{
Module B: NPS UI Function
}
